Refactor messaging files to simplify structure, set up conversation page
