+++ client/pages/messages/[id].js
@@ -1,9 +1,10 @@
-import { useState, useEffect } from 'react';
-import { useQuery } from 'react-query';
+import { useState, useEffect, useRef } from 'react';
+import { useQuery, useMutation } from 'react-query';
 import useAxios from '../../hooks/useAxios';
 
+import NavBar from '../../components/NavBar';
 import Container from '../../components/ui/Container';
-import Messages from '../../components/Messaging/Messages';
+import MessageList from '../../components/Messaging/Messages';
 
 export const getServerSideProps = async (ctx) => {
 	// TODO: Get the data from the server here using ctx.params.id
@@ -12,16 +13,58 @@ export const getServerSideProps = async (ctx) => {
 };
 
 function Conversation(props) {
-	const { isLoading, isError, data } = useQuery('conversations', () =>
-		useAxios({ url: `/conversations/${props.id}`, method: 'get' }),
+	const { isLoading, isError, data, refetch } = useQuery(
+		'conversations',
+		() => {
+			console.log('Ok Refreshing page');
+			return useAxios({ url: `/conversations/${props.id}`, method: 'get' });
+		},
+		{
+			refetchInterval: 10000,
+		},
+	);
+	console.log('Data >>>>', data);
+
+	const refreshButton = useRef(null);
+
+	const [text, setText] = useState('');
+
+	const mutation = useMutation((newMessage) =>
+		useAxios({ url: `/conversations/1`, method: 'post', params: newMessage }),
 	);
 
+	function handleSubmit() {
+		mutation.mutate({ body: text, sender_id: 1 });
+		setText('');
+		setTimeout(() => {
+			refreshButton.current.click();
+			console.log('Reload function >>>>> ', refreshButton.current);
+		}, 1500);
+	}
+
 	return (
-		<div className='flex flex-col justify-between'>
-			<Container title='Message'>
-				{/* <Messages key={data.id} messages={data.messages} /> */}
+		<>
+			<NavBar />
+			<Container title='Talk With Requester'>
+				{data && <MessageList key={data.id} {...data} />}
+				{/* COMPOSE MESSAGE */}
+				<div className='w-full flex justify-between'>
+					<textarea
+						className='flex-grow focus:bg-white m-2 py-2 px-4 mr-1 rounded-full border border-gray-300 bg-gray-200 resize-none'
+						rows='1'
+						placeholder='Message...'
+						onChange={(event) => setText(event.target.value)}
+						value={text}></textarea>
+					<button className='p-5' onClick={handleSubmit}>
+						Send
+					</button>
+				</div>
 			</Container>
-		</div>
+			<button ref={refreshButton} onClick={refetch}>
+				{' '}
+				Do refetch NOW{' '}
+			</button>
+		</>
 	);
 }
 //help request header
